refactor: move ModeSelector component to UI library directory
- Move ModeSelector component to src/components/ui/ directory
- Update import paths in all files using ModeSelector (OperationsPanel, tests, OrganizationPicker)
- Verify application builds correctly with type checking and linting
- Mark task T010 as completed in TODO.md
