+++ src/components/OrganizationPicker.tsx
@@ -1,6 +1,6 @@
 import { useState, useEffect, useCallback, useRef } from 'react';
 import Image from 'next/image';
-import { ActivityMode } from './ModeSelector';
+import { ActivityMode } from './ui/ModeSelector';
 import { useDebounceCallback } from '@/hooks/useDebounce';
 
 export type Organization = {
