+++ src/components/dashboard/OperationsPanel.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import ModeSelector, { ActivityMode } from '@/components/ModeSelector';
+import ModeSelector, { ActivityMode } from '@/components/ui/ModeSelector';
 import OrganizationPicker from '@/components/OrganizationPicker';
 import { DateRange, FilterState, Installation } from '@/types/dashboard';
 import { getGitHubAppInstallUrl } from '@/lib/dashboard-utils';
